removed some bulky code

+++ app/models.py
@@ -79,15 +79,11 @@ class Computer(models.Model):
     title = models.CharField(max_length=120)
     slug = models.SlugField(unique=True)
     description = models.TextField(blank=True, null=True)
-    price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    price = models.IntegerField( null=True, blank=True)
     image = models.ImageField(upload_to="posts/", blank=True, null=True)
     stock = models.PositiveIntegerField(null=True, blank=True)
     available = models.BooleanField(default=True)
-    compcategory = models.ForeignKey(
-        'CompCategory',
-        related_name="computers",
-        on_delete=models.CASCADE
-    )
+    compcategory = models.ForeignKey('CompCategory', related_name="computers", on_delete=models.CASCADE)
 
     def save(self, *args, **kwargs):
         self.slug = slugify(self.title)
